REFACTOR : Entity 속 ID의 데이터 타입 변경에 의한 수정(Integer->Long)

+++ src/main/java/com/example/deliveryapi/food/FoodJpaRepository.java
@@ -2,6 +2,6 @@ package com.example.deliveryapi.food;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FoodJpaRepository extends JpaRepository<Food, Integer> {
+public interface FoodJpaRepository extends JpaRepository<Food, Long> {
 
 }
